[RNE Rewrite] refactor(vad): rename streaming method, move op to utils/
- push -> detectVoiceOnStream (descriptive, pairs with detectVoice); its 'no
  transition' return is now VadEvent | undefined instead of | null.
- Move the extractFrames wrapper ops.ts -> extensions/speech/utils/vadUtils.ts.
  It's a model-specific op, not a domain-general one like cv's ops. Codify this
  in the add-native-extension skill: domain-general ops stay in ops.ts,
  model-/task-specific ops go under extensions/<domain>/utils/<name>.ts.
