+++ packages/react-native-executorch/src/extensions/speech/index.ts
@@ -1,1 +1,1 @@
-export * from './ops';
+export * from './utils/vadUtils';
